+++ exportplan/forms.py
@@ -42,7 +42,7 @@ class CountryDemographicsForm(forms.Form):
 
 class ExportPlanAboutYourBusinessForm(forms.Form):
     story = forms.CharField(
-        label='How we started',
+        label='How you started',
         required=False,
         widget=Textarea(attrs={
             'example': (
@@ -58,7 +58,7 @@ class ExportPlanAboutYourBusinessForm(forms.Form):
         ),
     )
     location = forms.CharField(
-        label="Where we're based",
+        label="Where you're based",
         required=False,
         widget=Textarea(attrs={
             'example': (
@@ -68,7 +68,7 @@ class ExportPlanAboutYourBusinessForm(forms.Form):
         ),
     )
     processes = forms.CharField(
-        label='How we make our products',
+        label='How you make your products',
         required=False,
         widget=Textarea(attrs={
             'example': (
@@ -79,25 +79,24 @@ class ExportPlanAboutYourBusinessForm(forms.Form):
         ),
     )
     packaging = forms.CharField(
-        label='Our packaging',
+        label='Your product packaging',
         required=False,
         widget=Textarea(attrs={
             'example': (
-                'Our packaging is themed around purity and comprises elegant clear glass bottles with '
-                'twist cap and unbleached paper label printed with natural dyes. Our distribution and '
-                'delivery boxes are 100% recycled cardboard. Each features our distinctive Dove label. '
+                'Our packaging is themed around purity and we use: <br> <li>elegant clear glass bottles</li>'
+                '<li>twist caps</li><li>unbleached paper label printed with natural dyes.</li>'
+                'Our distribution and delivery boxes are 100% recycled cardboard.'
+                'Each features our distinctive Dove label.'
             )}
         ),
     )
     performance = forms.ChoiceField(
-        label='Business performance',
+        label='Your business performance',
         choices=TURNOVER_CHOICES,
         required=False,
         widget=Select(attrs={
-            'example': (
-                'From 2015 to 2018 sales have grown on average 31% a year. Revenue flattened off '
-                'slightly in 2019 because of a UK distribution issue which has now been resolved. '
-                'We are on track to meet our sales targets for 2020.'
+            'description': (
+                'What is the annual turnover of your business?'
             ),
             'tooltip': (
                 'Give a summary of the past 3 to 5 years. You could include sales and profit figures '
@@ -328,31 +327,10 @@ class ExportPlanMarketingApproachForm(forms.Form):
                 'example liquor store email marketing campaigns.</p>'
             ),
             'description': (
-                '<p>Write down:</p><ul class="list-bullet"><li>what marketing you can do in-house</li>'
+                '<p class="m-0">To make sure you have the right amount of marketing resources make '
+                'sure you write down:</p><ul class="list-dot"><li>what marketing you can do in-house</li>'
                 '<li>the areas where your business will need support</li><li>external agencies or '
-                'people you’ll work with and what you expect them to do</li>'
+                'people you will work with and what they will be doing</li>'
             )}
-        ),
-    )
-    spending = forms.CharField(
-        label='How much do you think you’ll spend on marketing?',
-        required=False,
-        widget=NumberInput(attrs={
-            'example': (
-                '<p>We think we’ll spend around £13,000 a year in 2021 and 2022 on marketing for Dove Gin’s '
-                'launch in Australia.</p><p>The main costs will be:</p><ul><li>£500 per month for paid search '
-                'for a 6 month campaign late 2021</li><li>£5,000 agency fees</li></ul><p>Remaining budget tbc '
-                'but will need to include giveaways and potential cost of sponsoring launch nights at bars in '
-                'late 2021.</p>'
-            ),
-            'description': (
-                '<p>Explain how your total marketing spend breaks down.</p>'
-            ),
-            'placeholder': (
-                0.00
-            ),
-            'currency': (
-                'GBP'
-            )},
         ),
     )