+++ react-components/src/components/ProductFinder/CountryFinder.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import ReactDOM from 'react-dom'
 import ReactModal from 'react-modal'
 import PropTypes from 'prop-types'
@@ -9,15 +9,15 @@ import Confirmation from './MessageConfirmation'
 
 const suggested = [{'country':'France', 'region':'Europe'}, {'country': 'Spain', 'region':'Europe'}, {'country':'Italy', 'region':'Europe'}, {'country':'Jamaica',  'region':'Latin America and Caribbean'}]
 
-export function CountryFinder(props) {
+export function CountryFinderModal(props) {
   let modalContent
-  const { text } = props
-  const [modalIsOpen, setIsOpen] = useState(false)
-  const [selectedCountry, setSelectedCountry] = useState(text)
+  const { modalIsOpen, setIsOpen, text, addButton } = props
   const [countryList, setCountryList] = useState()
+  const [selectedCountry, setSelectedCountry] = useState(text)
   const [searchStr, setSearchStr] = useState()
   const [productConfirmationRequired, setProductConfirmationRequired] = useState(false)
   const [expandRegion, setExpandRegion] = useState(false)
+
 
   const openModal = () => {
     setProductConfirmationRequired(!!selectedCountry)
@@ -36,8 +36,10 @@ export function CountryFinder(props) {
   }
 
   const searchChange = (evt) => {
-    setSearchStr(evt.target.value.toUpperCase())
-    setExpandRegion(true)
+    if (evt.target.value.length > 0) {
+      setSearchStr(evt.target.value.toUpperCase())
+      setExpandRegion(true)
+    }
   }
 
   const toggleRegion = () => {
@@ -54,7 +56,7 @@ export function CountryFinder(props) {
         return null
       })
       setCountryList(regions)
-    })  
+    })
   }
 
   const modalAfterOpen = () => {
@@ -66,6 +68,7 @@ export function CountryFinder(props) {
 
   const saveCountry = (country) => {
     setSelectedCountry(country.name)
+
     let result = Services.updateExportPlan({
       export_countries: [
         {
@@ -92,7 +95,7 @@ export function CountryFinder(props) {
     })
   }
 
-  let regions = Object.keys(countryList || {}).map((region, index) => {
+  let regions = Object.keys(countryList || {}).sort().map((region, index) => {
     let countries = (countryList[region] || []).map((country, index) => {
       if (searchStr && country.name.toUpperCase().indexOf(searchStr) != 0) return ''
       return (
@@ -126,11 +129,13 @@ export function CountryFinder(props) {
   const buttonClass = `tag ${!selectedCountry ? 'tag--tertiary' : ''} tag--icon `
 
   return (
-    <span>
-      <button type="button" className={buttonClass} onClick={openModal}>
-        {selectedCountry || 'add country'}
-        <i className={`fa ${(selectedCountry ? 'fa-edit' : 'fa-plus')}`}/>
-      </button>
+    <>
+      { addButton &&
+        <button type="button" className={buttonClass} onClick={openModal}>
+          {selectedCountry || 'add country'}
+          <i className={`fa ${(selectedCountry ? 'fa-edit' : 'fa-plus')}`}/>
+        </button>
+      }
       <ReactModal
         isOpen={modalIsOpen}
         onRequestClose={closeModal}
@@ -168,7 +173,7 @@ export function CountryFinder(props) {
               <hr className="bg-black-70"/>
               <h3 className="h-s p-t-0">List of markets</h3>
               <p className="m-v-xs">
-                If you have an idea of where you want to export, choose from the list below. You can change this at any
+                If you have an idea of where you want to export, choose from the list below. <br/>You can change this at any
                 time.
               </p>
               <div className="grid">
@@ -183,14 +188,14 @@ export function CountryFinder(props) {
                   ></input>
                   <span className="visually-hidden">Search markets </span>
                   <i className="fas fa-search"></i>
-                  
+
                 </div>
               </div>
               <div className="grid">
                 <div className="c-full">
-                  <button type="button" key="region-expand" className="region-expand link" onClick={toggleRegion}>{expandRegion ? 'Collapse all' : 'Expand all' }</button>
+                  <button type="button" key="{index}" className="region-expand link" onClick={toggleRegion}>{expandRegion ? 'Collapse all' : 'Expand all' }</button>
                     <hr/>
-                  <ul className="country-list" onClick={selectCountry}>
+                  <ul className="country-list">
                     {regions}
                   </ul>
                 </div>
@@ -207,15 +212,42 @@ export function CountryFinder(props) {
         messageBody="if you've created an export plan, make sure you update it to reflect your new market. you can change target market at any time."
         messageButtonText="Got it"
       />
+    </>
+  )
+}
+
+CountryFinderModal.propTypes = {
+  addButton: PropTypes.bool,
+  modalIsOpen: PropTypes.bool,
+  setIsOpen: PropTypes.func.isRequired,
+  text: PropTypes.string
+}
+CountryFinderModal.defaultProps = {
+  addButton: true,
+  modalIsOpen: false,
+  text: '',
+}
+
+export const CountryFinder = ({ text }) => {
+
+  const [modalIsOpen, setIsOpen] = useState(false)
+
+  return (
+    <span>
+      <CountryFinderModal
+        text={text}
+        modalIsOpen={modalIsOpen}
+        setIsOpen={setIsOpen}
+      />
     </span>
   )
 }
 
 CountryFinder.propTypes = {
-  text: PropTypes.string
+  text: PropTypes.string,
 }
 CountryFinder.defaultProps = {
-  text: null
+  text: '',
 }
 
 export default function createCountryFinder({ ...params }) {