+++ react-components/src/components/RouteToMarket/index.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 
-import { RouteToMarketSection } from '@src/views/sections/MarketingApproach/RouteToMarket/RouteToMarketSection'
-import Services from '../../../../Services'
+import { RouteToMarketSection } from '@src/components/RouteToMarket/RouteToMarketSection'
+import Services from '../../Services'
 
 export const RouteToMarket = ({
   fields,
@@ -14,11 +14,14 @@ export const RouteToMarket = ({
 
   const addTable = () => {
     Services.createRouteToMarket({ ...formFields })
-      .then((data) => {
+      .then((data) => (
         setRoutes([
-          ...routes,
-          data
-        ])
+        ...routes,
+        data
+      ])))
+      .then(() => {
+        const newElement = document.getElementById(`Route to market ${routes.length+1}`).parentNode
+        newElement.scrollIntoView()
       })
       .catch(() => {})
   }
@@ -44,11 +47,19 @@ export const RouteToMarket = ({
 
   return (
     <>
-      {routes.length >=1 && routes.map((field) => RouteToMarketSection({ ...formData, update, deleteTable, field }))}
-      <div className='button--plus'>
-        <span className='icon--plus' />
-        <button type='button' onClick={addTable} className='button--stone'>Add route to market</button>
-      </div>
+      {routes.length >=1 && routes.map((field, id) => RouteToMarketSection({
+        ...formData,
+        data: formData.data.map((x) => x.name === 'route' ? { ...x, label: `${x.label} ${id+1}`} : x),
+        update,
+        deleteTable,
+        field,
+      }))}
+      <button
+        type='button'
+        className='button button--large button--icon'
+        onClick={addTable}>
+        <i className='fas fa-plus-circle' />Add route to market
+      </button>
     </>
   )
 }